Fixed dispatch function to be atomic
The whole function is critical, so it expects interrupts to be disabled
when entering. It delays restoring the interrupts until the very end of
the function.

NOTE: As per https://gcc.gnu.org/wiki/avr-gcc#Call-Used_Register we read
that the T-flag is considered call-used/clobbered.

There are two cases that a task can be resumed from:
 - The task is being run for the first time
 - The task is being resumed after being paused by kos_dispatch, which
   is a function call.

Since neither of these assume some state of the T flag, we clobber it in
order to restore the I flag at the end of dispatch (ret vs reti).

The dispatch function is never used to store/restore state for an ISR.
This is done using normal generated avr-gcc code. The ISR stack is
always the current task stack. For example, if an ISR interrupts on task
number 2 and calls a function that results in a context switch to task
number 1, task number 1 will immediately begin executing. Once it blocks
again, then task number 2 will be resumed at the point within the ISR
before the ISR finishes.

TODO: Make ISRs wait to call the dispatcher at the end via some sort of
isr enter and exit functions.