Fixed potential concurrency problem with mutex and isrs
First off, all of our functions that called "mutex exit" reenabled
interrupts no matter what. This is very bad, especially considering that
the AVR has certain interrupts which must not be nested (UART, anyone?).
These were all replaced with ATOMIC_BLOCK.

The other problem is that the scheduler was allowed to run halfway
through an interrupt. This could be bad since if the interrupt
interrupts a low priority task, the scheduler will immediately cause a
context switch when called partway through the interrupt. Not a big deal
for interrupts that do just one thing, but if the interrupt is a little
more complicated and calls multiple kos functions, this could be an
issue. I added the isr enter and exit functions to mitigate this by
delaying all scheduler calls to the end of an isr. Added some overhead.