Increased idle task size for ISR (35 + 17 > 48)
The ISR had a minimum stack requirement of 17 because of the call-used
registers (and we call a function, so it gets to store all of those)
plus a couple others and the isr return address. Obviously, this will
cause a problem when we attempt to dispatch from the idle task stack
after pushing 17 things. The problem in our case is that the semaphore
just happened to get corrupted to a 0x40 which caused weird stuff to
happen. Really hard to detect the bug...